Win: Fix scrollwheel issue on multi-monitor config
The LOWORD and HIWORD macros return an unsigned integer, but that
won't be correct if the viewer is being displayed on a monitor that is
above or to the left of the "main"/"primary" display, because at least
one of the coordinates will be negative in that case.

Fixes #69.

+++ win/vncviewer/ClientConnection.cpp
@@ -2528,8 +2528,8 @@ LRESULT CALLBACK ClientConnection::WndProc(HWND hwnd, UINT iMsg,
       SetFocus(hwnd);
 
       POINT coords;
-      coords.x = LOWORD(lParam);
-      coords.y = HIWORD(lParam);
+      coords.x = (LONG)(short)LOWORD(lParam);
+      coords.y = (LONG)(short)HIWORD(lParam);
 
       if (iMsg == WM_MOUSEWHEEL) {
         // Convert coordinates to position in our client area.
